Connect seek and tweak icons

+++ yamcs-studio-tycho/org.yamcs.studio.ui/src/main/java/org/yamcs/studio/ui/archive/ZoomSpec.java
@@ -37,7 +37,7 @@ public class ZoomSpec {
         return (int) Math.round((ms - startInstant) / pixelRatio);
     }
 
-    long convertPixelToInstant(int x) {
+    long convertPixelToInstant(long x) {
         return (long) (x * pixelRatio) + startInstant;
     }
